Successfully migrated pagination in separate component. Code is much cleaner.
