--- src/components/common/Navigation/Pagination/PaginationUI.jsx
@@ -1,1 +1,0 @@
-export const PaginationUI = () => {};
