+++ src/pages/Movies/Movies.jsx
@@ -16,25 +16,15 @@ import {
   selectSearchQuery,
   selectSearchTotalPages,
   setCategory,
-  setSearchPage,
   setSearchQuery,
 } from "../../utils/redux/searchSlice";
 
-import {
-  Button,
-  Container,
-  Content,
-  Header,
-  Pagination,
-  StyledLeftIcon,
-  StyledLink,
-  StyledRightIcon,
-  Text,
-} from "./styled";
+import { Container, Content, Header, StyledLink } from "./styled";
 import { toMoviePage } from "../../routes";
 import { NoResult } from "../NoResult/NoResult";
 import { Error } from "../Error/Error";
 import { StyledLoader } from "../../components/common/StyledLoader/styled";
+import { PaginationMovies } from "../../components/common/Navigation/Pagination/PaginationMovies";
 
 export default function Movies() {
   const dispatch = useDispatch();
@@ -67,54 +57,6 @@ export default function Movies() {
     dispatch(setPage(1));
   }, [dispatch, category]);
 
-  const prevPageHandler = () => {
-    if (page !== 1) {
-      dispatch(setPage(page - 1));
-    }
-  };
-
-  const prevSearchPageHandler = () => {
-    if (searchPage !== 1) {
-      dispatch(setSearchPage(searchPage - 1));
-    }
-  };
-
-  const nextPageHandler = () => {
-    if (page !== totalPages) {
-      dispatch(setPage(page + 1));
-    }
-  };
-
-  const nextSearchPageHandler = () => {
-    if (searchPage !== searchTotalPages) {
-      dispatch(setSearchPage(searchPage + 1));
-    }
-  };
-
-  const firstPageHandler = () => {
-    if (page !== 1) {
-      dispatch(setPage(1));
-    }
-  };
-
-  const firstSearchPageHandler = () => {
-    if (searchPage !== 1) {
-      dispatch(setSearchPage(1));
-    }
-  };
-
-  const lastPageHandler = () => {
-    if (page !== totalPages) {
-      dispatch(setPage(totalPages));
-    }
-  };
-
-  const lastSearchPageHandler = () => {
-    if (searchPage !== searchTotalPages) {
-      dispatch(setSearchPage(searchTotalPages));
-    }
-  };
-
   return (
     <Container>
       {status !== "error" && (
@@ -144,93 +86,13 @@ export default function Movies() {
                       </StyledLink>
                     ))}
                   </Content>
-                  <Pagination>
-                    {page && searchPage === 1 ? (
-                      <>
-                        <Button disabled>
-                          <StyledLeftIcon />
-                          <StyledLeftIcon />
-                          <p>First</p>
-                        </Button>
-                        <Button disabled>
-                          <StyledLeftIcon />
-                          <p>Previous</p>
-                        </Button>
-                      </>
-                    ) : (
-                      <>
-                        <Button
-                          onClick={
-                            !searchQuery
-                              ? firstPageHandler
-                              : firstSearchPageHandler
-                          }
-                        >
-                          <StyledLeftIcon />
-                          <StyledLeftIcon />
-                          <p>First</p>
-                        </Button>
-                        <Button
-                          onClick={
-                            !searchQuery
-                              ? prevPageHandler
-                              : prevSearchPageHandler
-                          }
-                        >
-                          <StyledLeftIcon />
-                          <p>Previous</p>
-                        </Button>
-                      </>
-                    )}
-                    {searchQuery ? (
-                      <span>
-                        <Text>Page</Text> <strong>{searchPage}</strong>{" "}
-                        <Text>of</Text> <strong>{searchTotalPages}</strong>
-                      </span>
-                    ) : (
-                      <span>
-                        <Text>Page</Text> <strong>{page}</strong>{" "}
-                        <Text>of</Text> <strong>{totalPages}</strong>
-                      </span>
-                    )}
-                    {page === totalPages && searchPage === searchTotalPages ? (
-                      <>
-                        <Button disabled>
-                          <p>Next</p>
-                          <StyledRightIcon disabled />
-                        </Button>
-                        <Button disabled>
-                          <p>Last</p>
-                          <StyledRightIcon />
-                          <StyledRightIcon />
-                        </Button>
-                      </>
-                    ) : (
-                      <>
-                        <Button
-                          onClick={
-                            !searchQuery
-                              ? nextPageHandler
-                              : nextSearchPageHandler
-                          }
-                        >
-                          <p>Next</p>
-                          <StyledRightIcon />
-                        </Button>
-                        <Button
-                          onClick={
-                            !searchQuery
-                              ? lastPageHandler
-                              : lastSearchPageHandler
-                          }
-                        >
-                          <p>Last</p>
-                          <StyledRightIcon />
-                          <StyledRightIcon />
-                        </Button>
-                      </>
-                    )}
-                  </Pagination>
+                  <PaginationMovies
+                    page={page}
+                    totalPages={totalPages}
+                    searchTotalPages={searchTotalPages}
+                    searchPage={searchPage}
+                    searchQuery={searchQuery}
+                  />
                 </>
               ) : (
                 <NoResult />
