+++ src/pages/People/People.jsx
@@ -1,12 +1,4 @@
 import { useDispatch, useSelector } from "react-redux";
-
-import {
-  Button,
-  Pagination,
-  StyledLeftIcon,
-  StyledRightIcon,
-  Text,
-} from "../Movies/styled";
 import { Container, Content, Header } from "./styled";
 import {
   fetchPopularPeople,
@@ -23,7 +15,6 @@ import {
   selectSearchQuery,
   selectSearchTotalPages,
   setCategory,
-  setSearchPage,
   setSearchQuery,
 } from "../../utils/redux/searchSlice";
 import { nanoid } from "nanoid";
@@ -32,6 +23,7 @@ import { NoResult } from "../NoResult/NoResult";
 import { Error } from "../Error/Error";
 import { StyledLoader } from "../../components/common/StyledLoader/styled";
 import { useEffect } from "react";
+import { PaginationPeople } from "../../components/common/Navigation/Pagination/PaginationPeople";
 
 export default function People() {
   const dispatch = useDispatch();
@@ -64,54 +56,6 @@ export default function People() {
     dispatch(setPage(1));
   }, [dispatch, category]);
 
-  const prevPageHandler = () => {
-    if (page !== 1) {
-      dispatch(setPage(page - 1));
-    }
-  };
-
-  const prevSearchPageHandler = () => {
-    if (searchPage !== 1) {
-      dispatch(setSearchPage(searchPage - 1));
-    }
-  };
-
-  const nextPageHandler = () => {
-    if (page !== totalPages) {
-      dispatch(setPage(page + 1));
-    }
-  };
-
-  const nextSearchPageHandler = () => {
-    if (searchPage !== searchTotalPages) {
-      dispatch(setSearchPage(searchPage + 1));
-    }
-  };
-
-  const firstPageHandler = () => {
-    if (page !== 1) {
-      dispatch(setPage(1));
-    }
-  };
-
-  const firstSearchPageHandler = () => {
-    if (searchPage !== 1) {
-      dispatch(setSearchPage(1));
-    }
-  };
-
-  const lastPageHandler = () => {
-    if (page !== totalPages) {
-      dispatch(setPage(totalPages));
-    }
-  };
-
-  const lastSearchPageHandler = () => {
-    if (searchPage !== searchTotalPages) {
-      dispatch(setSearchPage(searchTotalPages));
-    }
-  };
-
   return (
     <Container>
       {status !== "error" && (
@@ -136,93 +80,13 @@ export default function People() {
                       <PersonTile key={nanoid()} member={member} />
                     ))}
                   </Content>
-                  <Pagination>
-                    {page && searchPage === 1 ? (
-                      <>
-                        <Button disabled>
-                          <StyledLeftIcon />
-                          <StyledLeftIcon />
-                          <p>First</p>
-                        </Button>
-                        <Button disabled>
-                          <StyledLeftIcon />
-                          <p>Previous</p>
-                        </Button>
-                      </>
-                    ) : (
-                      <>
-                        <Button
-                          onClick={
-                            !searchQuery
-                              ? firstPageHandler
-                              : firstSearchPageHandler
-                          }
-                        >
-                          <StyledLeftIcon />
-                          <StyledLeftIcon />
-                          <p>First</p>
-                        </Button>
-                        <Button
-                          onClick={
-                            !searchQuery
-                              ? prevPageHandler
-                              : prevSearchPageHandler
-                          }
-                        >
-                          <StyledLeftIcon />
-                          <p>Previous</p>
-                        </Button>
-                      </>
-                    )}
-                    {searchQuery ? (
-                      <span>
-                        <Text>Page</Text> <strong>{searchPage}</strong>{" "}
-                        <Text>of</Text> <strong>{searchTotalPages}</strong>
-                      </span>
-                    ) : (
-                      <span>
-                        <Text>Page</Text> <strong>{page}</strong>{" "}
-                        <Text>of</Text> <strong>{totalPages}</strong>
-                      </span>
-                    )}
-                    {page === totalPages && searchPage === searchTotalPages ? (
-                      <>
-                        <Button disabled>
-                          <p>Next</p>
-                          <StyledRightIcon disabled />
-                        </Button>
-                        <Button disabled>
-                          <p>Last</p>
-                          <StyledRightIcon />
-                          <StyledRightIcon />
-                        </Button>
-                      </>
-                    ) : (
-                      <>
-                        <Button
-                          onClick={
-                            !searchQuery
-                              ? nextPageHandler
-                              : nextSearchPageHandler
-                          }
-                        >
-                          <p>Next</p>
-                          <StyledRightIcon />
-                        </Button>
-                        <Button
-                          onClick={
-                            !searchQuery
-                              ? lastPageHandler
-                              : lastSearchPageHandler
-                          }
-                        >
-                          <p>Last</p>
-                          <StyledRightIcon />
-                          <StyledRightIcon />
-                        </Button>
-                      </>
-                    )}
-                  </Pagination>
+                  <PaginationPeople
+                    page={page}
+                    totalPages={totalPages}
+                    searchPage={searchPage}
+                    searchTotalPages={searchTotalPages}
+                    searchQuery={searchQuery}
+                  />
                 </>
               ) : (
                 <NoResult />
